refactored the app to serve single user (me)
