--- src/routes/dashboard/+page.svelte
@@ -1,6 +1,0 @@
-<script>
-    import Savings from '$lib/Savings.svelte';
-    export let data
-</script>
-
-<Savings data={data.filteredData}/>
