--- src/routes/exchange_token/+page.svelte
@@ -1,7 +1,0 @@
-<script>
-    import { onMount } from 'svelte'
-
-    onMount(() => {
-        window.location = '/dashboard'
-    })
-</script>
